requirements.txt
Django==2.2.17
djangorestframework==3.12.2
django-rest-swagger==2.2.0